+++ src/lib/data/fetchers/index.ts
@@ -7,8 +7,10 @@
 export { BaseFetcher } from './base';
 export type { FetcherConfig, FetchResult } from './base';
 
-// Individual fetchers will be added here as they're implemented:
-// export { BitnodesFetcher } from './bitnodes';
+// Implemented fetchers
+export { BitnodesFetcher, getBitnodesFetcher } from './bitnodes';
+
+// TODO: Implement these fetchers
 // export { SolanaBeachFetcher } from './solanabeach';
 // export { GitHubFetcher } from './github';
 // export { EthernodesFetcher } from './ethernodes';